Move RealDiskInterface class to disk_interface.h.
This is a TODO in src/ninja_jumble.cc
